Map the jsdoc loaded modules to a test/mocked module
Unfortunately, jest is not using webpack to process it’s stuff (which
is a strong argument for using Karma instead of Jest), so the jsdoc
loader can’t/won’t work when using jest. Here, we’re configuring jest
to detect when we use the jsdoc loader and to provide a mock module.
This is not as good as just allowing it to work and do its thing in
tests, but it’s better than nothing.