Turn off the import lint rule for extraneous dependencies
Too many false positives because of our unusual structure. We have a
/src director which is what this project exports to users. However, we
also have a /lib directory with a documentation site built into it.
It’s not expected to be used directly by users, and therefore its
dependencies are dev dependencies. This rule thinks that those should
not be dev dependencies.